using load() ajax method to get the links without leaving page
For each headline in the blog, we create an empty target, insert it after the headline, and then establish a click handler.

The click handler finds the hyperlink within the headline, modifies the url to insert a space before the # (our page was set up specifically for this purpose -- the original version would take the user to that page, scrolled to that anchor. The anchor is inside a div with the same value as its id, so jQuery can use the same # to pull out just that div's content.

Because of the hyperlink in the original code, we prevent the default behavior to avoid actually going to the blog page now that we can retrieve it via Ajax. (Note that even though we caught the event in the hyperlink's container tag, the entire bubbling process takes place before the default behavior, and it doesn't matter which element we actually use to invoke preventDefault).