use cache and do ajax call only if it is empty
Prior Issue: we were loading the JSON every time the user changes their selection. 
Goal: My next goal was to rewrite the code so that we only make the request once, and then use a cached response when the user changes their choice in the select. Solution: The handling of the response has been split off into a standalone function, handleResponse, so that we can invoke it from several places. I also created a variable to hold the response data object.

How it works: When the select changes, we look to see if the response data object has been populated, and, if so, use it by invoking handleResponse. If it isn't there, then we make the Ajax call, and use handleResponse in our callback.