jQuery UI autocomplete, redirecting to the sections of the page, animation
I iterate ($('h2').each) over all h2s on the page and, for each one, push a new hash onto the titles array, with fields populated from the text (.html()) and id attribute (.attr('id')) of the respective h2.

With the titles array now populated appropriately, I use it as the source for the jQuery UI autocomplete on the #q input field. I specify the action resulting from the user selecting any of the suggested items with the select event handler, redirecting the user to the appropriate anchor (window.location.href = '#'+ui.item.url).

I use animate to change the background and text color of each row in the #fruits table, changing to a dark red background with white text when moused over, and back to white background with black text when moused out.